Django/NoUgly/requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
colorama==0.4.4
cryptography==36.0.1
defusedxml==0.7.1
Django==3.2.8
django-allauth==0.47.0
django-filter==21.1
django-phonenumber-field==6.0.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
httpie==3.0.2
idna==3.3
importlib-metadata==4.10.1
Markdown==3.3.6
multidict==6.0.2
mysqlclient==2.1.0
oauthlib==3.1.1
phonenumbers==8.12.41
Pillow==9.0.0
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyJWT==1.7.1
PySocks==1.7.1
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.8
wincertstore==0.2
zipp==3.7.0
